+++ src/capacitor-run.ts
@@ -20,16 +20,20 @@ import { Project } from './project';
 export function capacitorRun(project: Project, platform: CapacitorPlatform): string {
   let preop = '';
   let rebuilt = false;
+  let noSync = false;
 
   // If the user modified something in the editor then its likely they need to rebuild the app before running
   if (ionicState.projectDirty) {
     const channel = getOutputChannel();
     channel.appendLine('[Ionic] Rebuilding as you changed your project...');
     preop = ionicBuild(project) + ' && ';
+    preop += `npx cap copy ${platform} && `;
     rebuilt = true;
   } else {
     preop = preflightNPMCheck(project);
   }
+
+  noSync = ionicState.syncDone.includes(platform);
 
   ionicState.refreshDebugDevices = true;
   ionicState.lastRun = platform;
@@ -41,7 +45,7 @@ export function capacitorRun(project: Project, platform: CapacitorPlatform): str
     case MonoRepoType.yarn:
     case MonoRepoType.lerna:
     case MonoRepoType.npm:
-      return preop + capRun(platform, project.repoType, rebuilt);
+      return preop + capRun(platform, project.repoType, rebuilt, noSync);
     case MonoRepoType.nx:
       return preop + nxRun(project, platform);
     default:
@@ -50,11 +54,18 @@ export function capacitorRun(project: Project, platform: CapacitorPlatform): str
 }
 
 export function capacitorDevicesCommand(platform: CapacitorPlatform): string {
-  const ionic = exists('@ionic/cli') ? 'ionic ' : '';
+  const ionic = useIonicCLI() ? 'ionic ' : '';
   return `npx ${ionic}cap run ${platform} --list`;
 }
 
-function capRun(platform: CapacitorPlatform, repoType: MonoRepoType, noBuild: boolean): string {
+export function useIonicCLI(): boolean {
+  if (exists('@capacitor/cli')) {
+    return false;
+  }
+  return exists('@ionic/cli');
+}
+
+function capRun(platform: CapacitorPlatform, repoType: MonoRepoType, noBuild: boolean, noSync: boolean): string {
   const liveReload = vscode.workspace.getConfiguration('ionic').get('liveReload');
   const externalIP = vscode.workspace.getConfiguration('ionic').get('externalAddress');
   const prod: boolean = vscode.workspace.getConfiguration('ionic').get('buildForProduction');
@@ -69,7 +80,7 @@ function capRun(platform: CapacitorPlatform, repoType: MonoRepoType, noBuild: bo
     );
   }
 
-  const ionic = exists('@ionic/cli') ? 'ionic ' : '';
+  const ionic = useIonicCLI() || liveReload ? 'ionic ' : '';
 
   if (externalIP) {
     if (capRunFlags.length >= 0) capRunFlags += ' ';
@@ -80,10 +91,17 @@ function capRun(platform: CapacitorPlatform, repoType: MonoRepoType, noBuild: bo
     if (capRunFlags.length >= 0) capRunFlags += ' ';
     capRunFlags += '--prod';
   }
+
   if (noBuild) {
     if (capRunFlags.length >= 0) capRunFlags += ' ';
     capRunFlags += '--no-build';
   }
+
+  if (noSync) {
+    if (capRunFlags.length >= 0) capRunFlags += ' ';
+    capRunFlags += '--no-sync';
+  }
+
   capRunFlags += getConfigurationArgs();
 
   const pre =