+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Display the build configuration used for Angular projects where DefaultConfiguration is set
 - Fix for Capacitor Sync where non-default build configuration is used
+- Show Capacitor commands even if Cordova is still present
 
 ### Version 1.15.0