+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Display the build configuration used for Angular projects where DefaultConfiguration is set
 - Run speed optimization: if a project is changed and rebuilt make sure cap run does not build again
+- Optimization: if dependencies do not change then avoid cap sync when re-running
 - Fix for Capacitor Sync where non-default build configuration is used
 - Show Capacitor commands even if Cordova is still present